Week3/assignment/requirements.txt
# Core LangChain and LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0

# Structured output validation
pydantic>=2.0.0

# Prompt file parsing
pyyaml>=6.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Redis for production rate limiting/cost tracking
# redis>=5.0.0
